+++ src/withComponents/VideoToggle/VideoToggle.tsx
@@ -2,14 +2,16 @@ import React from 'react';
 import clsx from 'clsx';
 
 import Tooltip from '@material-ui/core/Tooltip';
+import { motion } from 'framer-motion';
 
 import { ReactComponent as CameraOn } from '../../images/icons/camera_on.svg';
 import { ReactComponent as CameraOff } from '../../images/icons/camera_off.svg';
 
 import useLocalVideoToggle from '../../hooks/useLocalVideoToggle/useLocalVideoToggle';
 
-import styles from './VideoToggle.module.css';
+import { videoToggleTransition, videoToggleSwitchVariants } from './VideoToggle-animation';
 
+import styles from './VideoToggle.module.css';
 interface IVideoToggleProps {
   compact?: boolean;
   border?: boolean;
@@ -24,30 +26,42 @@ export const VideoToggle: React.FC<IVideoToggleProps> = ({ compact = false, bord
       placement="bottom"
       PopperProps={{ disablePortal: true }}
     >
-      <div
-        className={clsx('u-cursorPointer u-flex', styles.toggleButtonBackdrop, {
+      <motion.div
+        className={clsx('u-cursorPointer u-flex u-positionRelative', styles.toggleButtonBackdrop, {
           uflexJustifyEnd: isVideoEnabled,
           [styles['toggleButtonBackdrop-border']]: border,
         })}
         onClick={toggleVideoEnabled}
+        animate={isVideoEnabled ? (compact ? 'onCompact' : 'on') : 'off'}
+        transition={videoToggleTransition}
       >
+        <motion.div className={styles.toggleSwitch} variants={videoToggleSwitchVariants}></motion.div>
         <div
-          className={clsx('u-flex u-flexJustifyCenter u-flexAlignItemsCenter', styles.toggleOption, styles.camOff, {
-            [styles.toggleOptionSelected]: !isVideoEnabled,
-            [styles.compactOption]: isVideoEnabled && compact,
-          })}
+          className={clsx(
+            'u-flex u-flexJustifyCenter u-flexAlignItemsCenter u-layerSurfaceBeta',
+            styles.toggleOption,
+            styles.camOff,
+            {
+              [styles.toggleOptionSelected]: !isVideoEnabled,
+              [styles.compactOption]: isVideoEnabled && compact,
+            }
+          )}
         >
           {isVideoEnabled && compact ? '•' : <CameraOff />}
         </div>
         <div
-          className={clsx('u-flex u-flexJustifyCenter u-flexAlignItemsCenter', styles.toggleOption, {
-            [styles.toggleOptionSelected]: isVideoEnabled,
-            [styles.compactOption]: !isVideoEnabled && compact,
-          })}
+          className={clsx(
+            'u-flex u-flexJustifyCenter u-flexAlignItemsCenter u-layerSurfaceBeta ',
+            styles.toggleOption,
+            {
+              [styles.toggleOptionSelected]: isVideoEnabled,
+              [styles.compactOption]: !isVideoEnabled && compact,
+            }
+          )}
         >
           {!isVideoEnabled && compact ? '•' : <CameraOn />}
         </div>
-      </div>
+      </motion.div>
     </Tooltip>
   );
 };